(PC-15471) recommended offer ids: do not fetch endpoint when user is not connected
	- useHomeRecommendedHits: extract hooks in separate files
	- improve tests
