+++ src/features/home/api/tests/useHomeRecommendedHits.test.ts
@@ -1,84 +1,63 @@
-import { renderHook, cleanup } from '@testing-library/react-hooks'
-import { rest } from 'msw'
-
-import { RecommendationParametersFields } from 'features/home/contentful'
-import * as AlgoliaModule from 'libs/algolia/fetchAlgolia/fetchAlgolia'
-import { mockedAlgoliaResponse } from 'libs/algolia/mockedResponses/mockedAlgoliaResponse'
-import { env } from 'libs/environment'
-import { queryCache, reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
-import { server } from 'tests/server'
-import { waitFor } from 'tests/utils'
+import { renderHook } from '@testing-library/react-hooks'
+import { UseMutationResult } from 'react-query'
 
 import {
-  useHomeRecommendedHits,
   getRecommendationEndpoint,
   getRecommendationParameters,
-} from '../useHomeRecommendedHits'
+  useHomeRecommendedHits,
+} from 'features/home/api/useHomeRecommendedHits'
+import { RecommendationParametersFields } from 'features/home/contentful'
+import { env } from 'libs/environment'
 
-const mockUserId = 30
-jest.mock('features/profile/api', () => ({
-  useUserProfileInfo: jest.fn(() => ({ data: { id: mockUserId } })),
-}))
-jest.mock('features/auth/settings')
+import * as algoliaRecommendedHitsAPI from '../useAlgoliaRecommendedHits'
+import * as recommendedIdsAPI from '../useHomeRecommendedIdsMutation'
 
-const objectIds = mockedAlgoliaResponse.hits.map(({ objectID }) => objectID)
-
-const fetchHits = jest.fn().mockResolvedValue(mockedAlgoliaResponse.hits)
-jest.spyOn(AlgoliaModule, 'fetchAlgoliaHits').mockImplementation(fetchHits)
-
-const endpoint = getRecommendationEndpoint(mockUserId, null) as string
+const mockUserId = 1234
+const position = {
+  latitude: 6,
+  longitude: 22,
+}
+const mockModuleId = 'abcd'
 
 describe('useHomeRecommendedHits', () => {
-  beforeEach(() => {
-    server.use(
-      rest.post(endpoint, (_req, res, ctx) =>
-        res(ctx.status(200), ctx.json({ playlist_recommended_offers: objectIds }))
-      )
-    )
+  const mutate = jest.fn()
+  jest
+    .spyOn(recommendedIdsAPI, 'useHomeRecommendedIdsMutation')
+    .mockReturnValue({ mutate } as unknown as UseMutationResult<
+      recommendedIdsAPI.RecommendedIdsResponse,
+      unknown,
+      recommendedIdsAPI.RecommendedIdsRequest,
+      unknown
+    >)
+
+  const algoliaSpy = jest
+    .spyOn(algoliaRecommendedHitsAPI, 'useAlgoliaRecommendedHits')
+    .mockImplementation(jest.fn())
+
+  it('should not call recommendation mutation when user is not connected', () => {
+    renderHook(() => useHomeRecommendedHits(undefined, position, mockModuleId))
+    expect(mutate).not.toHaveBeenCalled()
   })
 
-  afterEach(async () => {
-    await cleanup()
+  it('should call recommendation mutation when user is connected', () => {
+    renderHook(() => useHomeRecommendedHits(mockUserId, position, mockModuleId))
+    expect(mutate).toHaveBeenCalled()
   })
 
-  it('should not make any call if there is no recommendation module', async () => {
-    renderHook(() => useHomeRecommendedHits(undefined, null, 'abcd'), {
-      // eslint-disable-next-line local-rules/no-react-query-provider-hoc
-      wrapper: ({ children }) => reactQueryProviderHOC(children),
-    })
-
-    await waitFor(() => {
-      expect(fetchHits).not.toHaveBeenCalled()
-      expect(queryCache.get('recommendationOfferIds')).toBeUndefined()
-      expect(queryCache.get('recommendationHits')).toBeUndefined()
-    })
-  })
-
-  it('calls fetchAlgolia with params and returns data', async () => {
-    const { result } = renderHook(() => useHomeRecommendedHits(mockUserId, null, 'abcd'), {
-      // eslint-disable-next-line local-rules/no-react-query-provider-hoc
-      wrapper: ({ children }) => reactQueryProviderHOC(children),
-    })
-    const isUserUnderage = false
-    await waitFor(() => {
-      expect(result.current).toHaveLength(4)
-      expect(fetchHits).toHaveBeenCalledTimes(1)
-      expect(fetchHits).toHaveBeenCalledWith(objectIds, isUserUnderage)
-    })
+  it('should call algolia hook', () => {
+    renderHook(() => useHomeRecommendedHits(undefined, position, mockModuleId))
+    expect(algoliaSpy).toHaveBeenCalled()
+    renderHook(() => useHomeRecommendedHits(mockUserId, position, mockModuleId))
+    expect(algoliaSpy).toHaveBeenCalled()
   })
 })
 
 describe('getRecommendationEndpoint', () => {
-  it('should return undefined if no user id is provided', () => {
+  it('should return undefined when no user id is provided', () => {
     const endpoint = getRecommendationEndpoint(undefined, null)
     expect(endpoint).toBeUndefined()
   })
-
-  it('should return endpoint with latitude and longitude query params if position is provided', () => {
-    const position = {
-      latitude: 6,
-      longitude: 22,
-    }
+  it('should return endpoint with latitude and longitude query params when position is provided', () => {
     const endpoint = getRecommendationEndpoint(mockUserId, position)
     expect(endpoint).toEqual(
       `${env.RECOMMENDATION_ENDPOINT}/playlist_recommendation/${mockUserId}?token=${env.RECOMMENDATION_TOKEN}&longitude=${position.longitude}&latitude=${position.latitude}`
@@ -92,7 +71,7 @@ describe('getRecommendationParameters', () => {
     expect(result).toEqual({})
   })
 
-  it('should return parameters with mapped categories if provided', () => {
+  it('should return parameters with mapped categories when parameters are provided', () => {
     const parameters: RecommendationParametersFields = {
       title: 'some parameters',
       categories: ['Cinéma', 'Cours, ateliers', 'Livres'],
