+++ src/features/home/api/useHomeRecommendedHits.ts
@@ -1,51 +1,16 @@
-import { useEffect, useMemo, useState } from 'react'
-import { useMutation, useQuery } from 'react-query'
+import { useEffect, useState } from 'react'
 
-import { RecommendationParametersFields } from 'features/home/contentful'
-import { useIsUserUnderage } from 'features/profile/utils'
 import {
-  fetchAlgoliaHits,
-  filterAlgoliaHit,
-  useTransformAlgoliaHits,
-} from 'libs/algolia/fetchAlgolia'
+  RecommendedIdsRequest,
+  useHomeRecommendedIdsMutation,
+} from 'features/home/api/useHomeRecommendedIdsMutation'
+import { RecommendationParametersFields } from 'features/home/contentful'
 import { env } from 'libs/environment'
 import { GeoCoordinates } from 'libs/geolocation'
-import { eventMonitoring } from 'libs/monitoring'
-import { QueryKeys } from 'libs/queryKeys'
-import { IncompleteSearchHit, SearchHit } from 'libs/search'
+import { SearchHit } from 'libs/search'
 import { getCategoriesFacetFilters } from 'libs/search/utils'
 
-interface RecommendedIdsResponse {
-  playlist_recommended_offers: string[]
-}
-
-interface RecommendedIdsRequest {
-  start_date?: string
-  end_date?: string
-  isEvent?: boolean
-  categories?: string[]
-  price_max?: number
-}
-
-export const useHomeRecommendedHits = (
-  userId: number | undefined,
-  position: GeoCoordinates | null,
-  moduleId: string,
-  recommendationParameters?: RecommendationParametersFields
-): SearchHit[] | undefined => {
-  const recommendationEndpoint = getRecommendationEndpoint(userId, position) as string
-  const [recommendedIds, setRecommendedIds] = useState<string[]>()
-  const { mutate: getRecommendedIds } = useHomeRecommendedIdsMutation(recommendationEndpoint)
-
-  useEffect(() => {
-    const requestParameters = getRecommendationParameters(recommendationParameters)
-    getRecommendedIds(requestParameters, {
-      onSuccess: (response) => setRecommendedIds(response.playlist_recommended_offers),
-    })
-  }, [getRecommendedIds, recommendationParameters])
-
-  return useAlgoliaRecommendedHits(recommendedIds || [], moduleId)
-}
+import { useAlgoliaRecommendedHits } from './useAlgoliaRecommendedHits'
 
 export const getRecommendationEndpoint = (
   userId: number | undefined,
@@ -53,57 +18,8 @@ export const getRecommendationEndpoint = (
 ): string | undefined => {
   if (!userId) return undefined
   const endpoint = `${env.RECOMMENDATION_ENDPOINT}/playlist_recommendation/${userId}?token=${env.RECOMMENDATION_TOKEN}`
-
   if (position) return `${endpoint}&longitude=${position.longitude}&latitude=${position.latitude}`
   return endpoint
-}
-
-export const useHomeRecommendedIdsMutation = (recommendationUrl: string) => {
-  return useMutation(
-    QueryKeys.RECOMMENDATION_OFFER_IDS,
-    async (parameters: RecommendedIdsRequest) => {
-      try {
-        const response = await fetch(recommendationUrl, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            Accept: 'application/json',
-          },
-          body: JSON.stringify(parameters),
-        })
-        if (!response.ok) {
-          throw new Error('Failed to fetch recommendation')
-        }
-        const responseBody: RecommendedIdsResponse = await response.json()
-        return responseBody
-      } catch (err) {
-        eventMonitoring.captureException(new Error('Error with recommendation endpoint'), {
-          extra: { url: recommendationUrl },
-        })
-        return { playlist_recommended_offers: [] }
-      }
-    }
-  )
-}
-
-const useAlgoliaRecommendedHits = (ids: string[], moduleId: string): SearchHit[] | undefined => {
-  const isUserUnderage = useIsUserUnderage()
-  const transformHits = useTransformAlgoliaHits()
-
-  const moduleQueryKey = moduleId
-  const { data: hits } = useQuery(
-    [QueryKeys.RECOMMENDATION_HITS, moduleQueryKey],
-    async () => await fetchAlgoliaHits(ids, isUserUnderage),
-    { enabled: ids.length > 0 }
-  )
-
-  return useMemo(() => {
-    if (!hits) return
-
-    return (hits as IncompleteSearchHit[])
-      .filter(filterAlgoliaHit)
-      .map(transformHits) as SearchHit[]
-  }, [hits, transformHits])
 }
 
 export function getRecommendationParameters(
@@ -117,4 +33,25 @@ export function getRecommendationParameters(
     price_max: parameters?.isFree ? 0 : parameters?.priceMax,
     start_date: parameters?.beginningDatetime,
   }
+}
+
+export const useHomeRecommendedHits = (
+  userId: number | undefined,
+  position: GeoCoordinates | null,
+  moduleId: string,
+  recommendationParameters?: RecommendationParametersFields
+): SearchHit[] | undefined => {
+  const recommendationEndpoint = getRecommendationEndpoint(userId, position) as string
+  const [recommendedIds, setRecommendedIds] = useState<string[]>()
+  const { mutate: getRecommendedIds } = useHomeRecommendedIdsMutation(recommendationEndpoint)
+
+  useEffect(() => {
+    if (!recommendationEndpoint) return
+    const requestParameters = getRecommendationParameters(recommendationParameters)
+    getRecommendedIds(requestParameters, {
+      onSuccess: (response) => setRecommendedIds(response.playlist_recommended_offers),
+    })
+  }, [getRecommendedIds, recommendationParameters, recommendationEndpoint])
+
+  return useAlgoliaRecommendedHits(recommendedIds || [], moduleId)
 }
